Add docker-compose file MySQL Database
This will allow for a database to run on the development machines so
that there is something that can be used to test against, but it does
not run the actual backend in Docker itself. From initial poking around
it seems Go will handle that for us using `go run` for the different
entry points. As long as there isn't OS specific code, it should
function fine.

If there does end up being Host or OS specific code, then I'll expand
the docker-compose file to include a Development container as well. For
now I just want to avoid adding developmental complexity if it's
possible.